LayoutTests/platform/mac/accessibility/aria-invalid-expected.txt

   
This tests that aria-invalid causes the right attribute to be returned and it ensures a notification is sent when it changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.stringAttributeValue('AXInvalid') is 'true'
PASS accessibilityController.focusedElement.stringAttributeValue('AXInvalid') is 'grammar'
PASS accessibilityController.focusedElement.stringAttributeValue('AXInvalid') is 'false'
PASS accessibilityController.focusedElement.stringAttributeValue('AXInvalid') is 'false'
PASS addedNotification is true
PASS successfullyParsed is true

TEST COMPLETE
Notification received successfully.
